enricher/requirements.txt
kafka-python==2.0.2
redis==5.0.1
APScheduler==3.10.4
pandas==2.1.3
clickhouse-driver==0.2.6
